rename: hook 폴더명 변경 및 불필요 파일 제거
