+++ src/containers/button/ConfirmButtonContainer.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import axios from 'axios';
 import ConfirmButton from 'components/card/ConfirmButton';
-import { useAppState } from 'hooks/hooks';
+import { useAppState } from 'hooks';
 import { removePathnameCardEdit } from 'utils';
 
 function ConfirmButtonContainer() {
