+++ src/containers/button/NextButtonContainer.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ConfirmButton from 'components/card/ConfirmButton';
 import { INPUT } from '../../constants';
-import { useAppDispatch, useAppState } from 'hooks/hooks';
+import { useAppDispatch, useAppState } from 'hooks';
 import { createAction } from 'context/Provider';
 import { ActionType } from 'types';
