+++ src/containers/card/CardContainer.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react';
 import axios from 'axios';
 import { createAction } from 'context/Provider';
-import { useAppDispatch, useAppState } from 'hooks/hooks';
+import { useAppDispatch, useAppState } from 'hooks';
 import { ActionType } from 'types';
 import Card from 'components/card/Card';
 import { removePathnameCardEdit } from 'utils';
