+++ src/containers/modal/TypeButtonModalContainer.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Modal from 'components/modal/Modal';
 import { createAction } from 'context/Provider';
-import { useAppDispatch } from 'hooks/hooks';
+import { useAppDispatch } from 'hooks';
 import { ActionType } from 'types';
 import TypeButtonContainer from './TypeButtonContainer';
 import ModalPortal from 'portal';
