+++ src/pages/AddCardPage.tsx
@@ -7,7 +7,7 @@ import TypeButtonModal from 'containers/modal/TypeButtonModalContainer';
 import BackButtonContainer from 'containers/button/BackButtonContainer';
 import PageTitle from 'components/navigater/PageTitle';
 
-import { useAppState } from 'hooks/hooks';
+import { useAppState } from 'hooks';
 
 import { MarginWrapper, PageWrapper, TitleWrapper } from './style';
