--- src/scss/_flex.scss
@@ -1,61 +1,0 @@
-.d-flex {
-  display: flex;
-}
-
-.justify-content-start {
-  justify-content: flex-start;
-}
-.justify-content-end {
-  justify-content: flex-end;
-}
-.justify-content-center {
-  justify-content: center;
-}
-.justify-content-between {
-  justify-content: space-between;
-}
-.justify-content-around {
-  justify-content: space-around;
-}
-.justify-content-evenly {
-  justify-content: space-evenly;
-}
-
-.align-items-start {
-  align-items: flex-start;
-}
-.align-items-end {
-  align-items: flex-end;
-}
-.align-items-center {
-  align-items: center;
-}
-.align-items-baseline {
-  align-items: baseline;
-}
-.align-items-stretch {
-  align-items: stretch;
-}
-
-.align-content-center {
-  align-content: center;
-}
-.align-content-start {
-  align-content: flex-start;
-}
-.align-content-end {
-  align-content: flex-end;
-}
-.align-content-between {
-  align-content: space-between;
-}
-.align-content-around {
-  align-content: space-around;
-}
-.align-content-evenly {
-  align-content: space-around;
-}
-
-.flex-grow-1 {
-  flex-grow: 1;
-}
