--- src/scss/_typhography.scss
@@ -1,3 +1,0 @@
-.text-center {
-  text-align: center;
-}
