--- src/scss/style.scss
@@ -1,4 +1,0 @@
-@import './spacing';
-@import './typhography';
-@import './flex';
-@import './components/helptip';
